Changed target pose to correct points
Moved target pose from being directly on the April Tag to being a little bit in front of it and to the right or left depending on whether the right joystick is up or down. When the bot is outside of big R, it doesn't curve it just stays on the trajectory.