Added second half of 1DTD
In execute, added a lot of if statements to make the robot go on railroads once we reach PP (to a certain tolerance that we'll need to tune. Also I think we need to tune PP anyway). This is mainly done with the isAtPP boolean in 1DTD. Committing so that logic can be tested on simulation on a different computer.