comp day 1
another unsuccessful day for autons. progress is negligible, prooblems that were solved only gave way to new problems that are too inconsistent to understand or debug at comp. the only positive revelation is that pathfinding i guess is never going to be good for precise things but on the fly generation seems to be able to correct error between waypoints instead of at the end of a path which would help with the jittering.